requirements.txt
# Langchain Core
langchain
langchain-core

# OpenAI Integration
langchain-openai
openai

# Environment Variable Management
python-dotenv

# Streamlit for UI
streamlit

# Hive metastore client
hmsclient
# To connect HS2
pyhive
thrift_sasl

# To connect AWS
boto3